safety-rules: don't build all workspace binaries
