+++ consensus/safety-rules/Cargo.toml
@@ -25,7 +25,6 @@ libra-types = { path = "../../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
 serde = { version = "1.0.114", default-features = false }
 thiserror = "1.0.20"
-workspace-builder = { path = "../../common/workspace-builder", version = "0.1.0", optional = true }
 
 [dev-dependencies]
 criterion = "0.3.3"
@@ -33,14 +32,17 @@ tempfile = "3.1.0"
 consensus-types = { path = "../consensus-types", version = "0.1.0", features = ["fuzzing"] }
 libra-config = { path = "../../config", version = "0.1.0", features = ["fuzzing"] }
 libra-secure-storage = { path = "../../secure/storage", version = "0.1.0", features = ["testing"] }
-workspace-builder = { path = "../../common/workspace-builder", version = "0.1.0" }
 
 [[bench]]
 name = "safety_rules"
 harness = false
 required-features = ["testing"]
 
+[[test]]
+name = "safety-rules-testsuite"
+required-features = ["testing"]
+
 [features]
 default = []
 fuzzing = ["consensus-types/fuzzing", "libra-config/fuzzing"]
-testing = ["libra-secure-storage/testing", "workspace-builder"]
+testing = ["libra-secure-storage/testing"]
