+++ consensus/safety-rules/src/lib.rs
@@ -32,6 +32,3 @@ pub mod process_client_wrapper;
 #[cfg(any(test, feature = "testing"))]
 #[path = "test_utils.rs"]
 pub mod test_utils;
-
-#[cfg(test)]
-mod tests;
