+++ consensus/safety-rules/tests/safety-rules-testsuite/local.rs
@@ -1,9 +1,10 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{test_utils, tests::suite, SafetyRules};
+use crate::suite;
 use libra_crypto::{ed25519::Ed25519PrivateKey, Uniform};
 use libra_types::validator_signer::ValidatorSigner;
+use safety_rules::{test_utils, SafetyRulesManager};
 
 #[test]
 fn test() {
@@ -15,7 +16,9 @@ fn safety_rules(verify_vote_proposal_signature: bool) -> suite::Callback {
     Box::new(move || {
         let signer = ValidatorSigner::from_int(0);
         let storage = test_utils::test_storage(&signer);
-        let safety_rules = Box::new(SafetyRules::new(storage, verify_vote_proposal_signature));
+        let safety_rules_manager =
+            SafetyRulesManager::new_local(storage, verify_vote_proposal_signature);
+        let safety_rules = safety_rules_manager.client();
         (
             safety_rules,
             signer,
