+++ consensus/safety-rules/tests/safety-rules-testsuite/networking.rs
@@ -1,8 +1,8 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{test_utils, SafetyRulesManager};
 use libra_types::validator_signer::ValidatorSigner;
+use safety_rules::{test_utils, SafetyRulesManager};
 
 #[test]
 fn test_reconnect() {
