+++ consensus/safety-rules/tests/safety-rules-testsuite/vault.rs
@@ -1,10 +1,11 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{tests::suite, PersistentSafetyStorage, SafetyRulesManager};
+use crate::suite;
 use libra_crypto::{ed25519::Ed25519PrivateKey, Uniform};
 use libra_secure_storage::{KVStorage, Storage, VaultStorage};
 use libra_types::validator_signer::ValidatorSigner;
+use safety_rules::{PersistentSafetyStorage, SafetyRulesManager};
 
 /// A test for verifying VaultStorage properly supports the SafetyRule backend.  This test
 /// depends on running Vault, which can be done by using the provided docker run script in
@@ -24,7 +25,7 @@ fn safety_rules(verify_vote_proposal_signature: bool) -> suite::Callback {
         let mut storage = Storage::from(VaultStorage::new(host, token, None, None));
         storage.reset_and_clear().unwrap();
 
-        let waypoint = crate::test_utils::validator_signers_to_waypoint(&[&signer]);
+        let waypoint = safety_rules::test_utils::validator_signers_to_waypoint(&[&signer]);
         let storage = PersistentSafetyStorage::initialize(
             storage,
             signer.author(),
